+++ src/components/Table/components/AggregatedCellRenderer/AggregatedCellRenderer.tsx
@@ -10,7 +10,7 @@ import { CellType } from '@/types';
  */
 interface Props extends CellContext<unknown, unknown> {
   /**
-   * Bg Color
+   * Background Color
    *
    * @type {string}
    */