+++ src/components/Table/components/CellRenderer/components/DefaultCellRenderer/DefaultCellRenderer.tsx
@@ -31,7 +31,7 @@ interface Props {
   config: ColumnConfig;
 
   /**
-   * Bg Color
+   * Background Color
    *
    * @type {string}
    */