+++ src/components/editors/TablesEditor/components/TableEditor/components/ColumnsEditor/components/ColumnEditor/ColumnEditor.tsx
@@ -187,7 +187,7 @@ const columnPinDirectionOptions = [
   {
     value: ColumnPinDirection.LEFT,
     label: '',
-    icon: 'align-left',
+    icon: 'horizontal-align-left',
     description: 'Pinned to the left',
     ariaLabel: TEST_IDS.columnEditor.pinDirectionOption.selector(ColumnPinDirection.LEFT),
   },
@@ -199,7 +199,7 @@ const columnPinDirectionOptions = [
   {
     value: ColumnPinDirection.RIGHT,
     label: '',
-    icon: 'align-right',
+    icon: 'horizontal-align-right',
     description: 'Pinned to the right',
     ariaLabel: TEST_IDS.columnEditor.pinDirectionOption.selector(ColumnPinDirection.RIGHT),
   },
@@ -510,6 +510,19 @@ export const ColumnEditor: React.FC<Props> = ({ value, onChange, data, isAggrega
               {...TEST_IDS.columnEditor.fieldLabel.apply()}
             />
           </InlineField>
+          <InlineField label="Tooltip" grow={true}>
+            <Input
+              value={value.columnTooltip}
+              placeholder="Tooltip in column header"
+              onChange={(event) =>
+                onChange({
+                  ...value,
+                  columnTooltip: event.currentTarget.value,
+                })
+              }
+              {...TEST_IDS.columnEditor.fieldColumnTooltip.apply()}
+            />
+          </InlineField>
           <InlineFieldRow>
             <InlineField label="Font">
               <ColorEditor
@@ -706,19 +719,19 @@ export const ColumnEditor: React.FC<Props> = ({ value, onChange, data, isAggrega
                 options={[
                   {
                     value: ColumnAlignment.START,
-                    icon: 'align-left',
+                    icon: 'horizontal-align-left',
                     description: 'Start',
                     ariaLabel: TEST_IDS.columnEditor.fieldAppearanceAlignmentOption.selector(ColumnAlignment.START),
                   },
                   {
                     value: ColumnAlignment.CENTER,
-                    icon: 'bars',
+                    icon: 'horizontal-align-center',
                     description: 'Center',
                     ariaLabel: TEST_IDS.columnEditor.fieldAppearanceAlignmentOption.selector(ColumnAlignment.CENTER),
                   },
                   {
                     value: ColumnAlignment.END,
-                    icon: 'align-right',
+                    icon: 'horizontal-align-right',
                     description: 'End',
                     ariaLabel: TEST_IDS.columnEditor.fieldAppearanceAlignmentOption.selector(ColumnAlignment.END),
                   },