+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.3.0 (IN PROGRESS)
+
+### Features / Enhancements
+
+- Updated Highlighted row behavior (#250)
+- Updated align icons in panel options (#261)
+- Added tooltip option for column (#260)
+- Added contrast color for cell text if cell have data links (#253)
+
 ## 2.2.0 (2025-02-10)
 
 ### Features / Enhancements